+++ src/db/players.ts
@@ -1,0 +1,5 @@
+import { IPlayerWithPassword } from "../types/player.types";
+
+const players: IPlayerWithPassword[] = [];
+
+export default players;
