+++ src/db/rooms.ts
@@ -1,0 +1,5 @@
+import { room } from "../types/room.types";
+
+const rooms: room[] = [];
+
+export default rooms
