+++ src/types/battleshipWebSocket.type.ts
@@ -1,0 +1,8 @@
+import WebSocket from 'ws';
+
+interface IBattleshipWebSocket extends WebSocket {
+    playerIndex: string;
+    playerName: string;
+}
+
+export default IBattleshipWebSocket;
